+++ package.json
@@ -19,7 +19,7 @@
     "commit": "git-cz",
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "NODE_ENV=test react-scripts test --env=jest-environment-jsdom-sixteen --watchAll=false ----collectCoverage=true --detectOpenHandles",
+    "test": "react-scripts test --env=jest-environment-jsdom-sixteen --watchAll=false ----collectCoverage=true --detectOpenHandles",
     "lint": "eslint .",
     "fix-lint": "eslint . --fix",
     "ci": "npm run lint && npm run test",
@@ -39,35 +39,35 @@
     "not op_mini all"
   ],
   "dependencies": {
-    "@material-ui/core": "~4.11.3",
+    "@material-ui/core": "~4.12.1",
     "@material-ui/icons": "~4.11.2",
     "babel-polyfill": "~6.26.0",
     "font-awesome": "~4.7.0",
     "node-fetch": "~2.6.1",
-    "react": "~17.0.1",
-    "react-dom": "~17.0.1",
+    "react": "~17.0.2",
+    "react-dom": "~17.0.2",
     "react-scripts": "~4.0.3",
-    "react-share": "~4.3.1",
-    "recharts": "~2.0.7"
+    "react-share": "~4.4.0",
+    "recharts": "~2.0.9"
   },
   "devDependencies": {
-    "@commitlint/cli": "~12.0.0",
-    "@commitlint/config-conventional": "~12.0.0",
+    "@commitlint/cli": "~12.1.4",
+    "@commitlint/config-conventional": "~12.1.4",
     "@semantic-release/changelog": "~5.0.1",
     "@semantic-release/commit-analyzer": "~8.0.1",
     "@semantic-release/exec": "~5.0.0",
     "@semantic-release/git": "~9.0.0",
-    "@semantic-release/github": "~7.2.0",
-    "@semantic-release/npm": "~7.0.10",
-    "@semantic-release/release-notes-generator": "~9.0.1",
-    "@testing-library/jest-dom": "~5.11.9",
-    "@testing-library/react": "~11.2.5",
-    "@testing-library/user-event": "~12.7.3",
-    "commitizen": "~4.2.3",
+    "@semantic-release/github": "~7.2.3",
+    "@semantic-release/npm": "~7.1.3",
+    "@semantic-release/release-notes-generator": "~9.0.3",
+    "@testing-library/jest-dom": "~5.14.1",
+    "@testing-library/react": "~12.0.0",
+    "@testing-library/user-event": "~13.1.9",
+    "commitizen": "~4.2.4",
     "cz-conventional-changelog": "~3.3.0",
     "husky": "~4.3.8",
     "jest-environment-jsdom-sixteen": "~1.0.3",
-    "semantic-release": "~17.3.9"
+    "semantic-release": "~17.4.4"
   },
   "husky": {
     "hooks": {
